Extract markdown format method, fix list item (md)
